requirements.txt
fastapi==0.75.0
python-dotenv==0.10.1
requests==2.21.0
uvicorn==0.17.6